% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{foo}
\alias{foo}
\title{A Function to Foo Your Data}
\usage{
foo(x)
}
\arguments{
\item{x}{(numeric, length of 1) A vector to foo.}
}
\value{
Returns a vector the same length as x with a value of 1 added to it.
This is another sentence.
}
\description{
This is a description of my function
}
\details{
Here are some more details about \code{x}.
}
\examples{

#Here is a standard way to use the foo function.
foo(1)

#Here is how you apply foo to a vector.
foo(1:10)

}
