% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fooz.R
\name{fooz}
\alias{fooz}
\title{A function to Add two to x}
\usage{
fooz(x)
}
\arguments{
\item{x}{(numeric, length 1) a vector to foo}
}
\value{
This returns a vector that is the same length of x.
}
\description{
This takes a number and then adds two to it.
}
\details{
here are more details a lot of details so many details we have to
print on another line
}
